+++ RELEASE-NOTES.md
@@ -1,4 +1,32 @@
-## Version 3.0.1
+## Version 3.0.4: 
+
+* Removed popup menu for the browser action; now, clicking on 
+  the Tranquility icon will make the page readable. For the 
+  other actions provided through the popup menu, users will have 
+  to go to the options page or use the Read Later button after 
+  running tranquility.
+
+* Added a few "preset" configurations (color schemes and font colors) 
+  rather than having to customize each option manually. 
+  This is experimental at this time; will explore the option of 
+  allowing users to import a configuration file for this in the future.
+
+* Changed the background page to "about:blank" (instead of mozilla.org) 
+  when loading offline pages from the options window.
+
+* Minor bug fix to remove the progress bar which was not being removed 
+  in some corner cases.
+
+* Bug fix to correctly load original page when toggling Tranquility mode 
+  for an offline page.
+
+* Bug fix to handle loading mixed security content pages correctly
+
+* Bug fix to handle pre-formatted pages correctly 
+
+--------------------------------
+
+## Version 3.0.1:
 
 * Tranquility Reader 3.0.1 has been rewritten using WebExtensions APIs